Catching UIApplicationMain exceptions
Implemented a Ti build plugin and extended the module in
order to catch and report the description of exceptions in
UIApplicationMain.
The plugin adds some wrapping code to the main.m file automatically
generated by Titanium, which, when an exception is catch, records  its
description to NSUserDefaults. That value is picked up by HockeyApp at
the next app start, and it’s used for augmenting the crash report.